+++ src/driver/DriverFactory.ts
@@ -5,6 +5,7 @@ import {SqlServerDriver} from "./sqlserver/SqlServerDriver";
 import {OracleDriver} from "./oracle/OracleDriver";
 import {SqliteDriver} from "./sqlite/SqliteDriver";
 import {CordovaDriver} from "./cordova/CordovaDriver";
+import {SqljsDriver} from "./sqljs/SqljsDriver";
 import {MysqlDriver} from "./mysql/MysqlDriver";
 import {PostgresDriver} from "./postgres/PostgresDriver";
 import {Driver} from "./Driver";
@@ -31,6 +32,8 @@ export class DriverFactory {
                 return new SqliteDriver(connection);
             case "cordova":
                 return new CordovaDriver(connection);
+            case "sqljs":
+                return new SqljsDriver(connection);
             case "oracle":
                 return new OracleDriver(connection);
             case "mssql":